segments code into several new files
